Move dotcom e2e smoke tests to tests/smoke
Rename legacy dotcom e2e specs into tests/smoke and update imports to the new relative paths. Update e2e/README to document that legacy smoke specs live in tests/smoke, are separate from scenario tests, and do not run on CI. Adjust playwright.config.ts by adding a smokeTestMatch and setting the chromium project's testMatch to run the smoke suite (keeping scenarios separate). This preserves the reset-based legacy runner and clarifies migration guidance.
