+++ apps/dotcom/client/e2e/tests/smoke/database.spec.ts
@@ -1,8 +1,8 @@
-import { Editor } from '../fixtures/Editor'
-import { sleep } from '../fixtures/helpers'
-import { HomePage } from '../fixtures/HomePage'
-import { Sidebar } from '../fixtures/Sidebar'
-import { expect, test } from '../fixtures/tla-test'
+import { Editor } from '../../fixtures/Editor'
+import { sleep } from '../../fixtures/helpers'
+import { HomePage } from '../../fixtures/HomePage'
+import { Sidebar } from '../../fixtures/Sidebar'
+import { expect, test } from '../../fixtures/tla-test'
 
 test('correctly clears db ', async ({ page, browser, editor, sidebar, database }) => {
 	const documentsToAdd = 3
